🚧  (design-system) anything with `next` moves to custom
This is going to get real old with `*.stories`
 if we had to directly call other components.

This fixes a `__dirname` issue with vite/next
 since we are not technically loading or
 needing `next` for any component (which is good).

BREAKING CHANGE: 💥️ KBar moved to Custom Components
