+++ packages/design-system/src/components/Card/index.ts
@@ -1,12 +1,1 @@
 export { Card } from './Card'
-export {
-  css_card,
-  Card as CardShow,
-  CardContent,
-  CardImage,
-  CardImageContainer,
-  CardLink,
-  CardMeta,
-  CardTitle,
-} from './Show'
-export { CardSpotify, CardOuter, ImageBlur } from './Spotify'
