+++ packages/design-system/src/components/index.ts
@@ -27,7 +27,7 @@ export { BoxLink } from './BoxLink'
 export { Button, ButtonDemo, ButtonIcon, ButtonMarketing } from './Button'
 export { Callout } from './Callout'
 export { Caption } from './Caption'
-export { Card, CardOuter, CardSpotify, ImageBlur } from './Card'
+export { Card } from './Card'
 export {
   Carousel,
   CarouselArrowButton,
@@ -83,22 +83,6 @@ export { Heading } from './Heading'
 export { Icon } from './Icon'
 export { IconButton } from './IconButton'
 export { Image } from './Image'
-export {
-  // Native
-  KBarProvider,
-  // Wrapper Components
-  KBarDefault,
-  KBarPortal,
-  KBarSearch,
-  KBarSearchResult,
-  KBarSearchResults,
-  KBarFooter,
-  KBarSubscriptions,
-  // Wrapper Components Styles
-  KBarPositionerStyle,
-  KBarAnimatorStyle,
-  KBarSearchStyle,
-} from './KBar'
 export { Kbd } from './Kbd'
 export { Label } from './Label'
 export { Link } from './Link'
