+++ packages/design-system/src/custom/Hero/index.ts
