+++ packages/design-system/src/custom/KBar/KBar.styles.ts
